refactor:  moved ~30  lines out of the dems-engine service, dropping it from 461 to well below the 450 limit. Business logic is entirely unchanged.
